F1Bot (f1_bot v0.6.0)

API for retrieving data between different system components.

Functions

api_is_not_streaming?()

can_reload_session?()

demo_mode?()

Returns true if demo mode is enabled.

demo_mode_url()

Returns configured demo mode URL or nil if not configured.

driver_info_by_abbr(driver_abbr)

Same as driver_info_by_number/1 but takes the driver's 3-letter abbreviation

driver_info_by_number(driver_number)

Returns information about the driver with the given number such as their name, picture URL, team, and team color.

driver_list()

Returns the list of drivers in the current session.

driver_session_data(driver_number)

Returns full session data from the DriverDataRepo for the driver with the given number. Contains all laps, stints and other data collected about the driver in this session.

driver_summary(driver_no)

Returns the race summary of the driver with the given number which contains the driver's stints, top speed, best and average lap times, and best and average sector times.

fast_forward_replay(seconds)

Fast-forwards the session replay started with replay_session_live/2 by a given number of seconds.

fetch_env(key)

Same as get_env/2 but returns an ok tuple or :error.

get_env(key, default \\ nil)

Gets a configuration value for F1Bot application.

is_race?()

Returns an ok-wrapped boolean indicating whether the current session is a race or another type of session (FP/Q). Returns an error if the session info is not available.

lap_number()

Returns the current lap number.

reload_session(light_data, force_reload \\ false)

Fetches session archives from a given URL, silently replays them, and replaces the current session with the replayed session. Useful for testing and restoring data from the last session on server restart.

reload_session(url, light_data, force_reload)

reload_session(url, processing_opts, light_data, force_reload)

replay_session_live(playback_rate \\ 1)

Similar to reload_session/2, but replays the session in live, as if the session was currently in progress (equivalent to demo mode). Useful for development and debugging.

replay_session_live(url, playback_rate)

reset_session()

Resets the session state and clears all caches

resync_state_events()

Emits all known events from the current session state to the event bus to (1) update the delayed events caches (which are used on the website) and (2) synchronize the state of all connected Phoenix clients. This includes events such as session information, driver list, and session summaries for all drivers.

This should rarely (if ever) be needed as relevant events are automatically emitted on changes, which should keep the states synchronized at all times.

See F1Bot.F1Session.EventGenerator.StateSync for more details.

session_best_stats()

Returns the best lap time, best sector times, and top speed in this session.

session_clock()

Returns the current session clock - time remaining in qualifying and FP sessions or time left on the 2h timer during races.

session_copy(light_copy \\ true)

Returns a copy of the current F1Session session state for analysis.

light_copy controls whether "heavy" data, such as position and car telemetry data are stripped from the copy to reduce its size.

session_info()

Returns information about the session such as the session type (FP/Q/R) and name.

session_status()

Returns the current session status (e.g. started, in progress, ended)

stop_session_replay()

Stops the session replay started with replay_session_live/2.

sync_archive_cache(recheck_index \\ false)

Downloads all available session archives and caches them locally. You shouldn't need to call this function directly, as the archives are cached automatically when they are first accessed.

track_status_history()

Returns a history of track status changes (safety cars, red flags, etc.)