F1Bot.DelayedEvents (f1_bot v0.6.0)

Functions

available_delays()

clear_all_caches()

default_delay()

delay_step()

delayed_topic_for_event(scope, delay_ms)

fetch_latest_event(delay_ms, event_scope)

See F1Bot.DelayedEvents.Rebroadcaster.fetch_latest_event/2.

is_valid_delay?(delay_ms)

max_delay_ms()

min_delay_ms()

oneshot_init(scopes, delay_ms)

Send the latest event for each topic pair for topics that follow the init + delta pattern, e.g. charts where the init event contains the bulky chart specification and later events only contain new data points to add.

push_to_all(events)

subscribe_with_delay(scopes, delay_ms, send_init_events)