F1Bot.ExternalApi.Discord.Permissions (f1_bot v0.3.1)

Functions

everyone_has_external_emojis?(guild_id)

@spec everyone_has_external_emojis?(pos_integer()) ::
  {:ok, boolean()} | {:error, any()}

Checks whether @everyone role has "use external emojis" permission.

This also determines whether we can use external emojis in slash command responses due to Discord not respecting permissions granted via individual roles.

Discussion of this API limitation: https://github.com/discord/discord-api-docs/issues/2612

guild_role_permissions(guild_id, role_id)

@spec guild_role_permissions(pos_integer(), pos_integer()) ::
  {:ok, [Nostrum.Permission.t()]} | {:error, any()}