F1Bot.F1Session (f1_bot v0.4.0)

Holds all state related to a given F1 session and coordinates data processing across modules in F1Bot.F1Session scope.

All code in this scope is fully functional, without side effects. To communicate with other components that have side effects, such as posting to Twitter, it generates events that are processed by the caller, i.e. F1Bot.F1Session.Server.

Types

t()

@type t() :: %F1Bot.F1Session{
  clock: F1Bot.F1Session.Clock.t() | nil,
  driver_cache: F1Bot.F1Session.DriverCache.t(),
  driver_data_repo: F1Bot.F1Session.DriverDataRepo.t(),
  event_generator: F1Bot.F1Session.EventGenerator.t(),
  lap_counter: F1Bot.F1Session.LapCounter.t(),
  race_control: F1Bot.F1Session.RaceControl.t(),
  session_info: F1Bot.F1Session.SessionInfo.t(),
  session_status: atom() | nil,
  track_status_history: F1Bot.F1Session.TrackStatusHistory.t()
}

F1 Session State

Functions

driver_info_by_abbr(session, driver_abbr)

driver_info_by_number(session, driver_number)

driver_list(session)

driver_session_data(session, driver_number)

driver_summary(session, driver_number)

make_state_sync_events(session, local_time)

See F1Bot.F1Session.EventGenerator.make_state_sync_events/2.

new()

periodic_tick(session, local_time)

push_driver_list_update(session, drivers)

push_lap_counter_update(session, current_lap, total_laps, timestamp)

push_lap_number(session, driver_number, lap_number, timestamp)

push_lap_time(session, driver_number, lap_time, timestamp)

push_position(session, driver_number, position)

push_race_control_messages(session, messages)

push_sector_time(session, driver_number, sector, sector_time, timestamp)

push_session_info(session, session_info, local_time, ignore_reset)

push_session_status(session, session_status)

push_stint_data(session, driver_number, stint_data)

push_telemetry(session, driver_number, channels)

push_track_status(session, track_status, timestamp)

reset_session(session, local_time)

session_best_stats(session)

session_clock_from_local_time(session, local_time)

update_clock(session, server_time, local_time, remaining, is_running)