F1Bot.F1Session.DriverDataRepo.BestStats (f1_bot v0.5.1)

Processes and holds personal best and session-best statistics, such as fastest lap times, sector times, and top speed.

Types

fastest_sectors()

@type fastest_sectors() :: %{
  required(1) => pos_integer() | nil,
  required(2) => pos_integer() | nil,
  required(3) => pos_integer() | nil
}

t()

@type t() :: %F1Bot.F1Session.DriverDataRepo.BestStats{
  fastest_lap_ms: pos_integer() | nil,
  fastest_sectors: fastest_sectors(),
  personal_best: %{
    required(pos_integer()) =>
      F1Bot.F1Session.DriverDataRepo.PersonalBestStats.t()
  },
  top_speed: non_neg_integer() | nil
}

Session-wide stats for fastest laps, top speed across all drivers

Functions

new()

push_personal_best_stats(self, pb_stats)