F1Bot.F1Session.DriverDataRepo.DriverData (f1_bot v0.2.0)

Coordinates processing and stores data for a driver, e.g. laps, fastest lap, car telemetry, stint data and more.

Types

t()

@type t() :: %F1Bot.F1Session.DriverDataRepo.DriverData{
  current_lap_number: non_neg_integer(),
  fastest_lap: Timex.Duration.t(),
  laps: F1Bot.F1Session.DriverDataRepo.Laps.t(),
  number: pos_integer(),
  position_hist: F1Bot.F1Session.Common.TimeSeriesStore.t(),
  stints: F1Bot.F1Session.DriverDataRepo.Stints.t(),
  telemetry_hist: F1Bot.F1Session.Common.TimeSeriesStore.t(),
  top_speed: non_neg_integer(),
  top_speed_curr_lap: non_neg_integer()
}

Driver session data

Functions

light_copy(self)

Callback implementation for F1Bot.F1Session.LightCopy.light_copy/1.

new(number)

push_lap_number(self, lap_number, timestamp)

@spec push_lap_number(t(), pos_integer() | nil, DateTime.t()) :: t()

push_lap_time(self, lap_time, timestamp)

push_position(self, position)

push_sector_time(self, sector, sector_time, timestamp)

push_stint_data(self, stint_data)

push_telemetry(self, telemetry)