F1Bot.F1Session.DriverDataRepo.Laps (f1_bot v0.5.1)

Stores all information about laps driven by a driver in the current session, e.g. lap number, lap time and timestamp.

Types

t()

@type t() :: %F1Bot.F1Session.DriverDataRepo.Laps{
  data: %{required(pos_integer()) => F1Bot.F1Session.DriverDataRepo.Lap.t()}
}

Functions

fastest(laps)

fetch_by_number(laps, lap_no)

mark_outliers(laps, all_lap_times, around_ts \\ nil)

Determine outliers in the given laps based on lap time compared to all other laps that occurred around a similar time (+/- @outlier_window_ms). Lap is marked as an outlier if it's greater than factor * minimum lap time in the window.

new()

personal_best_stats(laps)

push_speed(self, current_lap_number, speed)

push_timing_data(self, current_lap_number, timing_data)