F1Bot.F1Session.Server (f1_bot v0.6.0)

GenServer that holds the live F1Bot.F1Session instance, acts as an entrypoint for incoming live timing packets and executes all side effects by passing event messages to F1Bot.Output.Discord and F1Bot.Output.Twitter via PubSub.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

driver_info_by_abbr(driver_abbr)

driver_info_by_number(driver_number)

driver_list()

driver_session_data(driver_number)

driver_summary(driver_no)

evaluate(session_fn)

Calls the given function with current F1Session as argument, the function should return the modified F1Session.

process_live_timing_packet(packet, processing_options)

process_transcript(transcript)

replace_session(session)

reset_session()

resync_state_events()

server_via()

session_best_stats()

session_clock_from_local_time(local_time)

session_info()

session_status()

set_local_time_mode(mode, extrapolation_rate \\ nil)

start_link(init_arg)

state(light_copy)

track_status_history()