F1Bot.F1Session.DriverDataRepo.BestStats (f1_bot v0.4.0)

Processes and holds best-of session statistics, such as the overall fastest lap and top speed.

Types

fastest_sectors()

@type fastest_sectors() :: %{
  required(1) => Timex.Duration.t() | nil,
  required(2) => Timex.Duration.t() | nil,
  required(3) => Timex.Duration.t() | nil
}

t()

@type t() :: %F1Bot.F1Session.DriverDataRepo.BestStats{
  fastest_lap: Timex.Duration.t(),
  fastest_sectors: fastest_sectors(),
  top_speed: non_neg_integer()
}

Session-wide stats for fastest laps, top speed across all drivers

Functions

new()

push_end_of_lap_result(self, eol_result)

Tracks session-wide best lap and top speed stats, and creates events when a driver beats a session-wide or personal best record.

push_end_of_sector_result(self, eos_result)

Tracks session-wide best sector times, and creates events when a driver beats a session-wide record.