F1Bot.F1Session.DriverDataRepo.BestStats (f1_bot v0.2.0)

Processes and holds best-of session statistics, such as the overall fastest lap and top speed.

Types

fastest_sectors()

@type fastest_sectors() :: %{
  required(1) => Timex.Duration.t() | nil,
  required(2) => Timex.Duration.t() | nil,
  required(3) => Timex.Duration.t() | nil
}

t()

@type t() :: %F1Bot.F1Session.DriverDataRepo.BestStats{
  fastest_lap: Timex.Duration.t(),
  fastest_sectors: fastest_sectors(),
  top_speed: non_neg_integer()
}

Session-wide stats for fastest laps, top speed across all drivers

Functions

new()

push_lap_time(self, lap_time)

push_sector_time(self, sector, sector_time)

push_top_speed(self, speed)