F1Bot.F1Session.Server (f1_bot v0.2.1)

GenServer that holds the live F1Bot.F1Session instance, acts as an entrypoint for incoming live timing packets and executes all side effects by passing event messages to F1Bot.Output.Discord and F1Bot.Output.Twitter via PubSub.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

driver_info_by_abbr(driver_abbr)

driver_info_by_number(driver_number)

driver_session_data(driver_number)

push_live_timing_packet(packet)

replace_session(session)

server_via()

session_info()

session_status()

start_link(init_arg)

state(light_copy)

track_status_history()